requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.0
python-socketio==5.10.0
python-dotenv==1.0.0
pydantic==2.4.2
